requirements.txt
psycopg2-binary
python-dateutil
pytest
pytest-mock
pandas
duckdb